requirements.txt
fastapi==0.124.4
uvicorn==0.41.0
pydantic==2.12.5
scikit-learn==1.8.0
numpy==2.4.2
ujson
slowapi==0.1.4
fastapi-cache2==0.2.2